@cronnoss 2023-10-24
Results of benchmark runs:

- before optimizations:
```
go test -v -count=1 -timeout=30s -tags bench .
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 306.961167ms / 300ms
    stats_optimization_test.go:47: memory used: 294Mb / 30Mb
```
- after optimizations:
```
go test -v -count=1 -timeout=30s -tags bench .
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 102.516459ms / 300ms
    stats_optimization_test.go:47: memory used: 3Mb / 30Mb